# event-v2-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample event service API",
        "title": "Event V2 Service API",
        "contact": {},
        "version": "2.0.0"
    },
    "host": "localhost:3003",
    "basePath": "/api/v2",
    "paths": {
        "/api/v2/events/email": {
            "post": {
                "description": "Process email event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Process email event",
                "parameters": [
                    {
                        "description": "Email Event",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EmailEvent"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Success response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v2/events/inventory": {
            "post": {
                "description": "Process inventory event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Process inventory event",
                "parameters": [
                    {
                        "description": "Inventory Event",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.InventoryEvent"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Success response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v2/events/payment": {
            "post": {
                "description": "Process payment event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Process payment event",
                "parameters": [
                    {
                        "description": "Payment Event",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PaymentEvent"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Success response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v2/health": {
            "get": {
                "description": "Get service health status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check endpoint",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "service": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "models.EmailEvent": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "send_at": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "models.InventoryEvent": {
            "type": "object",
            "properties": {
                "operation": {
                    "description": "\"add\" ou \"remove\"",
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "warehouse_id": {
                    "type": "string"
                }
            }
        },
        "models.PaymentEvent": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "payment_type": {
                    "type": "string"
                }
            }
        }
    }
}